rework read/writes and get rid of queue
Since there is no strict requirement to emit no 'data' event
if the stream is paused, it is possible to get rid of the queue
and simplify read/write logic.

NOTE: 'data' event is emitted duing addFile() so pipe()
(or 'data' listener) should be set up prior to that.

Also, emit actuall Error object in case of error.